Change from flyout to a new page according to latest policy. Add common functions to helper
